Added the ranking selector.
Since I can't assume that the population will be ordered,
(bc it's a List interface)
I can't sort it there, so I copy the List into an array and sort
it there. If it were an array list I could just do everything on the
list.
There is probably a nicer way to write that comparator.